// src/app/api/model/watchlist-out.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WatchlistItem } from './watchlist-item';


export interface WatchlistOut { 
    _id: string;
    userId: string;
    name: string;
    description?: string | null;
    tickers: Array<WatchlistItem>;
    createdAt: string;
    updatedAt: string;
    isDefault?: boolean;
}