// src/app/api/model/watchlist-create.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WatchlistCreate { 
    name: string;
    description?: string | null;
    tickers?: Array<string> | null;
    isDefault?: boolean | null;
}