// src/app/api/model/watchlist-update.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WatchlistUpdate { 
    name?: string | null;
    description?: string | null;
    isDefault?: boolean | null;
}